Add timestamp checker endpoint for Herika app
Herika Android app cannot know what the latest ts/gamets timestamps are in eventlog history
therefore it will always clash with existing DB unless the user carefully inputs timestamps
into the app which is problematic. Any mistake can lead to a desync and then either Herika app
or Skyrim mod will be getting the wrong context for responses due to messed up ordering.

This patch adds a simple http endpoint that checks the latest eventlog entry in DB
and returns ts and gamets in response. Herika app is updated to check that endpoint
every time before it sends inputtext or chat events. Now the Skyrim mod and Herika app
should never clash with each other even if running on an existing DB.

If the app does not detect the new endpoint, it will fall back to the previous way
of handling timestamps which is starting from 0 or starting from whatever was the last one
inputted by the user in app config (which is not recommended).